+++ neo/CMakeLists.txt
@@ -771,6 +771,7 @@ set(src_imgui
 	
 	sys/sys_imgui.h
 	sys/sys_imgui.cpp
+	sys/imgui_savestyle.cpp
 )
